Add two things (totally untested! so don't use yet):
An AutoUpdate class. Subclass to get an update() function that is called at the start of each update cycle. The eliminates the nuisance bigworld had where multiple framework pieces needed to rely on anyone using them to call update() or else they wouldn't work.

serial_runnable template, for classes that want to run a single async operation at a time. This class is a little different than the big world one, it strictly enforces allowing only a single running operation at a time, so clients can't swamp the system.