Implement picking for perspective sprites.
This is a pretty terrible implementation, and should be replaced by someone who knows what they're doing. It works by taking the screen point, walking all enabled items in the tree, and translating each item into screen space, then running hit detection. So performance can't be great. A bigger issue is that some sprites are in screen space and some in world -- so there's a hack function on Sprite (setIsInScreenCoordsHack()) that needs to be called if you want to receive touch events on a sprite in world space. This would generally only apply to the top-level sprite containing the world-coordinate sprites, I think.