Add a new sprite, NinePatch, used for drawing rounded rects, drop shadows, etc. Functions like the BigWorld RoundedRectSprite, which functioned like the Android 9 Patch. This one isn't quite done yet -- only the top 3 cells are currently being figured, still need to do the middle and bottom rows. Also need to finish autodetecting the format of the image, so you can feed it images with or without the extra border of sizing info.
The NinePatch makes use of a new construct, the ImageSource. ImageSources are a way to abstract where images come from. Currently there's just a single source for loading images from a file, but hopefully I'll have a chance to add another one for generating circular images, which will be used for the drop shadow effects. The ImageSource pattern is really how images should be handled all through the system, including the existing Image sprite, so we could generate images for that, too.