Begin work on a flexible image generation system using the arc concept from my previous projects. Not complete yet; this is about 90% of the busy work, but missing the core image generation routines.
Also, introduce the notion of FrameworkResources -- resources that are compiled into the app. Not a great place for large amounts of image data, but perfect for the arc scripts that do things like render drop shadows for the NinePatch. From now on, apps should include FrameworkResources.rc in their projects.