Add a meta data utility, for querying image files for meta info. This replicates what's in Image sprite, so other classes can make use of it, and Image sprite should be changed to make use of it.
Also, found out that Image sprite, like the image loader, was using 32bpp surfaces when it had to fall back on image loading to get meta info, so BLEH. Fixed that. Actually, cinder probably has a way to even just load in a single channel, which might be worth looking into (and also, of course, optimizing for popular formats, and avoiding the whole thing altogether).