Add ability to scale PDF sprites.
Note: This has one flaw. The way I wanted to do it doesn't disrupt the sprite API at all, that is, scaling of the PDF is transparent to the outside world, it doesn't affect the size of the PDF sprite. The problem is that in the drawing, what I have to do to draw the full size, unscaled PDF into the scaled area does something funky to the drawing area, creating a flicker in any area where the texture isn't being drawn. This is kind fine, because it's written so that the texture always draws in the full area. The problem occurs when the PDF sprite is semi transparent, and you see some of that flickering. If I want to keep things working the way they are probably the only solution is to add a flag so that scaling isn't applied in the drawClient() function.