Refactor the engine touch receiving code, in preparation to support TUIO objects.
Also, I noticed something that I imagine has no impact in practice, but could be the source of an inefficiency: The touch processing was locking the touch queue the entire time all of the events were being handled by the app. Shouldn't be a big deal, but the ideal behaviour would be to pop events off the touch input queue in the lock, then handle them outside it, so that's what it now does.