Remove include path to GL from the video gstreamer 1.0 property sheets. Projects shouldn't be exposing any path but the one to themselves, and specifically with GL, it was causing an interference with some projects, leading to unresolved externals.
I'm not totally clear if the project does need a path to the GL; seems reasonable, but I took it out here and it was fine. So either it also exists in the project itself (which is where I believe it should be), or we should add it there if we need to (but, again, seems to compile and run fine in its current state).