Clean up the QueryResult object: Remove dependency on RecycleArray, add support for 64 bit ints. No functional changes.
Add PersistentCache object. Currently unused. Was going to use it for the ImageMetaData, but as I look at it, the way it needs to be implemented, there will be a global cache that will store metadata for every DsCinder application on the computer. Shouldn't be an issue in production, where there should generally only be one, but for devs it makes me a little nervous about performance. I think the solution is to expand the persistentcahce with the notion of a primary key on a string field, and use that to store data in an unordered_map instead of a vector. But that's not implemented yet.