Text: improve color rendering. SEE NOTE
Previously, the color for text was pulled from Cairo, unpremultiplied, multiplied with the current color and opacity, then again premultiplied with the alpha from Cairo. This created some color halo-ing around the edges of text, particularly noticeable on thinner text. The new method uses the text color and the alpha value from cairo, which produces slightly more bold fonts but with much cleaner colors. The downside is that if you're using <span> tags in the text to adjust the color of text, you'll need to use the old rendering style using PreserveSpanColors.

NOTE: if you've used span tags in text for color, you'll need to add preserve_span_colors="true" or mText->setPreserveSpanColors(true);  to enable the old color rendering model.