Sprite: add lambda to evaluate inner hit
The advantage of this is so you don't have to override the sprite type just to change the picking logic. For example, if you have an image button that's an odd shape and want to ignore taps to part of it (like a parallelogram or a hexagram or something), you can just apply a lambda function instead of making a whole custom sprite type.