This commit adds ds::ui::Grid, a CSS Grid Layout implementation described here: https://www.w3.org/TR/2013/WD-css3-grid-layout-20130402/#layout-algorithm
It's the result of weeks of work, squashed into a single commit for easier comparison. New properties and functions have been added to ds::ui::Sprite to make it work with grid layouts. Derived UI classes, like Text, Image and LayoutSprite, have been updated to work well when placed inside grid cells.

For now, we don't support named grid lines, so use numbers to specify in which column(s) or row(s) you want to place your sprites. Additionally, we don't support automatic placement of content, because the heuristics of a DS_Cinder app differ in many ways from web content.

Gaps are supported, as are intrinsics and most of the functions (e.g. "minmax" or "repeat") available to create your grid. Value units like "px", "%", "vw" and "vmin" are supported, but type units like "em" are not.

We hope this is useful and a welcome addition to DS_Cinder.